Added some major enhancements.
Made it so you no longer have to type the entire word of a promt to do what you want in the app. For example, instead of having to type the word extract to launch the pack extractor, you can just simply type the letter e, and it will essentially do the same thing.
Reduced the time it takes to load the program significantly. This program will no longer use virtual dialogs to reveal input prompts moving forward.